bufferqueue: Fix return value when no output is available
The old code returned 0 (which for a read function is supposed to mean
EOF) when it didn't know if it was a case of EOF or an EAGAIN. Let's
go with assuming EAGAIN, and see what happens.

Also, ensure that errno is set for functions that might not set it
themselves, to something vaguely relevant.

Change-Id: I8e81e5992533c8e1bba1549c3ccb2628ef57c620
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>