macros: Get rid of a lot of cruft
This patch removes the public declarations of a bunch of
functions that are only used internally by the macro machinery,
as well as marking those functions as "static" inside macros.c.
In order to avoid prototype hell in the C-file, we have to
rearrange the file a bit to make sure called functions appear
before their callers, so the patch looks bigger than it is.

While at it, we remove very nearly every non-threadsafe
function, which basically only called out to the threadsafe
one anyway and weren't used anywhere inside the core.

This is basic prep-work in order to start supporting macros
via modules, so that a module can define itself as macro
module and set functions in a struct, something like so:
  int macros.init();
  int macros.deinit();
  int macros.get(macro_context *ctx, char *key);
  int macros.set(macro_context *ctx, char *key, (char *)(*callback)...);
  int macros.unset(macro_context *ctx, char *key);
  void macros.set_host(macro_context *ctx, struct host *h);
  void macros.set_<object>(macro_context *ctx, struct <object> *obj);
  char *macros.process(macro_context *ctx, char *str);

Change-Id: Id6517f388b49952734d1a22047d7ae74722a9fde
Signed-off-by: Andreas Ericsson <ae@op5.se>