bufferqueue: Make _read not push -1 bytes and not leak memory
The buffer can be stack allocated quite well, and we'll thus only do a
one-time copy to get it into the bufferqueue, allocated as an
appropriately-sized block, wasting no memory.

The alternative, to use _push_block instead of _push in _read for a
heap-allocated block instead, would have worked just as well, wasting a
tiny bit of memory and saving a tiny bit of CPU, however, the
inbetween thingy that was there earlier doesn't work.

Change-Id: Ib4f27a63e979dbb5e3e57cfa66d8ddf580ca7fcc
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>