kvvec_ekvstr: Define wellformedness better
Be stricter in what we consider well-formed ekvstr's. In particular, the
shortest well defined ekvstr is "=", which should be interpreted as the
key "" (empty string) and value "" (empty string). More than one
reserved character ('=', ';') for a pair is not allowed. These should be
escaped if they are to be included in the key or value string.

Change-Id: I8b7b4b76dafc0aef0df972b84debde3229734843
Signed-off-by: Anton Lofgren <alofgren@op5.com>