Add worker tests
These tests are written in a scripting language aimed at
the dumbest script interpreter this world has ever seen.
Fortunately, it's enough to describe most of the ways
in which plugins can misbehave towards the reaping code.

The number of tests required for full coverage of all the
different ways that plugins can behave weirdly is so far
unknown, but adding them in the ridiculously simple script
language is easy, which means we can build on it as we go
along.

All that aside; The most important part about this patch is
probably the comment explaining why we sometimes really can't
get the output of jobs we run (which I've researched once
before, but forgot the details of and had to spend half a day
to look up again), but the tests themselves provide decent
coverage of the master<->worker ipc as well as the worker
output reaper.

Change-Id: Ie89a08d2edf75f33e5280fbf30a2cb29da263f0d
Signed-off-by: Andreas Ericsson <ae@op5.se>