Allow disabling status data updates altogether
A lot of systems no longer use the status.log file for anything,
since Naemon provides livestatus for that, which is a very superior
way of getting status data out of naemon. With this patch, one can
set status_update_interval=0 in order to get rid of writing the
status log altogether.

Change-Id: Ie4c1c14fc1baa410cf8fb192a2c4693fe8a0521f
Signed-off-by: Andreas Ericsson <ae@op5.se>