configuration: Make unwritable check_result_path non-critical
The check_result_path is a heritage from a much, much darker
age, when checkresults were being written to disk before being
parsed and handled. Since the core is no longer crazy, we don't
actually do that anymore, and check_result_path being unwritable
should, at most, be a warning and not an error.

This fixes issue #mon-7957.

Change-Id: I2f3a38734de1271fb1a8fab5d72c812fdbb5798c
Signed-off-by: Andreas Ericsson <ae@op5.se>