spec: Legacy init script/customisation migrations
We've historically recommended users to put stuff in their
/etc/profile.d to add environment settings to their nagios/monitor
process. The old custom init script was retired in the nagios->naemon
move in favour of the stock/upstream one and the stock one sources
/etc/init.d/functions which, among other things, (sanely) resets the
PATH variable. This is one notable instance of a number of variables
which users might have customised by dropping things into
/etc/profile.d. To ameliorate this issue, and make transitions smoother,
we now look for the most common known cases of such customisations and
build an /etc/sysconfig/naemon file which has the same effect (i.e, it
is sourced by the init script).

This file also contains the logic for creating ramdisk directories,
which was previously found in /etc/init.d/monitor.

/etc/sysconfig/monitor has been demoted to a symlink to
/etc/sysconfig/naemon for consistency, and the contents of the former
are now included in the latter for backwards compatibility.

Sanely enough, we do not touch /etc/sysconfig/naemon if it already
exists, since that would most likely introduce a ton of
unwanted/unexpected changes to the users system.

This resolves MON-7972 and MON-7975.

Change-Id: Ic08e04dedae09790e8997e4004422f77204af88d
Signed-off-by: Anton Lofgren <alofgren@op5.com>