init: Make use of existing sysvinit infrastructure
Reinventing this particular wheel has been done enough times by all
the different distros. Let's not do it again. This should be compatible
with redhat (and derivates), SuSE (and derivates), and debian (and
derivates).

This should fix MON-7940, since we now actually wait for Naemon to be
stopped before returning from a `service naemon stop`, or equivalent
invocation.

... And this is why systemd is a good idea.

Change-Id: I6c310649e79354163de126fd4e065261d9251d91
Signed-off-by: Anton Lofgren <alofgren@op5.com>