janitor: Don't destroy NULL hash tables
Doing so triggers a GLib assertion, which is ugly, even though we
(presumably validly) try to destroy these in some circumstances. For
example, when calling cleanup (and subsequently free_memory), when we
fail to claim the Naemon lockfile.

A bunch of our test cases also triggers this assertion, which makes it
impossible to build Naemon with NDEBUG.

This fixes the cosmetic part of MON-8039.

Change-Id: Ic460fae99089fe97a0e6e4a0ac1413ce18c7f4f5
Signed-off-by: Anton Lofgren <alofgren@op5.com>