janitor: DRY, return ERROR, don't exit
The cleanup()/exit() routine is done in the only callsite for this
function already. Let's minimize the number of calls to exit() and get
rid of redundant ones.

This makes the code a bit easier to follow and also gives us more
consistent logging for daemonisation errors ("Bailing out due to failure
to daemonize....").

Change-Id: I70f37caa289027a96507d68004c72e8537881106
Signed-off-by: Anton Lofgren <alofgren@op5.com>