xod: Warn when same-name services are assigned via hostgroups
With commit 436dc7a01f707fe89ba4b40cf9b6446597a3b6f2, we began
to make sure that services on hostgroups never override services
assigned directly to hosts, but in the process the warning that
tells a user that multiple same-name services are assigned to a
host if both (well, all) of those services come from hostgroups.

This patch reintroduces that warning. It's only a warning though,
since configuration still passes when it happens.

Change-Id: I8ec52d9435b398962734a010765576fce5ef19f1
Signed-off-by: Andreas Ericsson <ae@op5.se>