spec: Additional sysconfig/naemon tweaking
Just like the check_oracle stuff handled in f15cdc80, we need to handle
the PERLLIB environment variable found in /etc/profile.d/op5plugins.sh.

The old init script executed nagios via a "su - <user>" call, which
includes the target user's environment, hence including the scripts in
/etc/profile.d/. The new init script doesn't. This env var won't be
exported when naemon is started on-boot. But it will be exported when a
sysadmin manually executes "mon restart" via some interactive shell.

We said that we didn't want to ever touch /etc/sysconfig/naemon if it
was already in place. But users are already running f15cdc80, and we
have already created /etc/sysconfig/naemon on their behalf, without the
PERLLIB stuff. So now we need to modify the file anyway. This patch
makes sure that we only modify the file if we haven't already.

As we always install /etc/profile.d/op5plugins.sh, no questions asked,
forcing this into /etc/sysconfig/naemon is not a big difference imo. If
a user *really* don't want naemon started with this env var, they could
just unset it below our code, or put our code in a subshell, or w/e.

This resolves the last/new part of MON-7972.

Change-Id: Ieb2580b4a85e8ecb74c0e80ffdcb74b250ab79ec
Signed-off-by: Philip Eklöf <peklof@op5.com>