service api: Don't crash on NULL names
The "find_<object>()" functions are part of our exported object
API, and as such it shouldn't crash when it's fed bad input. This
patch makes us avoid that by checking our input and return NULL
early on in case it's invalid.

This makes the API behave the same way now as it did before we
moved to using g_hash, and since at least parts of our own code
relies on that behaviour and handles a NULL return as an error,
we should reintroduce it.

This fixes MON-8127

Change-Id: Ice9cb937708c12a63cd20f3c7da4760cbfce3971
Signed-off-by: Andreas Ericsson <ae@op5.se>