checks: Use pristine options in check handlers
It's possible for a monitored object's options to be overwritten during
execution of the check handler. This means that using the object's
options is something one rarely would want to do, since they might
have changed since the event was first triggered.

In particular, this fixes a bug where the CHECK_OPTION_FORCE_EXECUTION
flag was lost during execution of the handler, which lead to the check
never actually being forced but rather had to go through the usual
constraint controls. This made it impossible to execute checks in a
number of scenarios, such as (among other things) checks being disabled
either globally or for the specific object.

To fix this, the check handlers now use the pristine check options, as
they were on function entry.

This fixes MON-8108.

Change-Id: Ie772a05bc705b6089538f91e8fde234758f88dae
Signed-off-by: Anton Lofgren <alofgren@op5.com>