naemo: Don't close stdout and stderr until we're done daemonizing
Doing so prevents error messages from being displayed, which is
pretty bad in case we fail to daemonize because the disk is full,
in which case nothing will be written to the logfile, so stdout
is all we have to work with.

This patch makes sure the error message is displayed to the user
but doesn't actually fix any other problem. It's a continuation
of the fix for MON-8117.

Change-Id: I8931500bc229c389e290a640da4ac2aceb0ebe51
Signed-off-by: Andreas Ericsson <ae@op5.se>