Resconstruct downtime depth from retention data
Previously, we'd jump through several hoops to make sure that we
triggered the same code paths post-restart as pre-restart in the
downtime handling, by setting downtime struct members *just* the right
way.

Reconstructing this information is definitely not the responsibility of
the downtime code, it's the responsibility of the retention data reader.

This fixes MON-8490

Change-Id: I7885eb22d9fbc77efadadd076d74809c1450591d
Signed-off-by: Anton Lofgren <alofgren@op5.com>