init: Fix a race in the stop() function on RedHat
Specifying the -b flag here is all kinds of wrong, since it's not used
for specifying the executable of the daemon (it's used for matching
against the "exe" entry in the proc fs when finding the pid of a running
process). By using it here, we're causing the killproc function to
offset its arguments to interpret our actual executable as the signal to
send to the PID found in the pid-file, which in turn makes it not do
synchronous shutdown (i.e it doesn't wait for the Naemon process to
actually finish).

This leads to a race which is mostly notable when doing "service naemon
stop; service naemon start" in quick succession (aka "service naemon restart").

To prevent this from creeping back in, this patch adds a couple of more
scenarios (restart, reload, etc.) to the ci_config.yml file.

This fixes MON-8566

Change-Id: I90819b5b3ff1867ac05ab2bbc884ee1e9ccfb5f8
Signed-off-by: Anton Lofgren <alofgren@op5.com>