Stop spamming logs if check_result_path isn't there
It's a deprecated feature and a lot of systems won't ever use it,
so convert the log message into a debug one, and enhance it with
the reason. We enhance the log message from the config parsing in
the same way, except we leave that at the normal level, as it's
only logged once per config-parsing, which can't really be
considered "spamming" by even the most retarded standards.

We also mark the configuration variables associated with
processing check results from disk as deprecated, although
we still allow them to be set. We no longer ship them in the
sample config, however, so hopefully people will never really
see them.

Resolves MON-8381.

Change-Id: I772f8b42b0dfa57e2e33fb44025498c65e22c959
Signed-off-by: Andreas Ericsson <ae@op5.se>