perfdata: Invalidate closed perfdata FD
This patch fixes a race condition that might occur when we've released
a file descriptor that is subsequently reused by another thread.

In particular, the problem manifests by a deadlock from a sequence of
events such as:

1 Naemon opens a perfdata file and an associated file descriptor
2 Naemon writes some perfdata to that file and closes the file,
 effectively releasing the file descriptor, without actually
 invalidating the file descriptor
3 Livestatus opens a socket, and is allocated the same file descriptor
 that Naemon just released and calls select on it
4 Naemon Receives a check result, enters the performance data recording
 routine and checks that the file descriptor is valid (by comparing it >= 0)
5 Naemon tries to flush its buffer queue to the file described by the FD, which
 is now the Livestatus socket.
6 Everything deadlocks

To guard against regressions of this bug, this patch also makes sure to
log all failed attempts to flush the perfdata buffer when a seemingly
valid FD is held.

This fixes MON-8868.

Change-Id: Ic00eaf877facb869d4343358a6ea27b89e88a710
Signed-off-by: Anton Lofgren <alofgren@op5.com>