query-handler: Log IO-broker error when socket registration fails
Previously we'd only log whatever strerror(errno) gave us, which
is 0 (=Success) when the issue is related to how the library is
used.

This patch rectifies that, giving us a decent chance of finding
out what the actual problem is by looking at the logfiles. It
has no functional impact, however, and has no issue related to
it.

Change-Id: I925e1e7d70031bf9beb4096dcbafd7cc10970e5c
Signed-off-by: Andreas Ericsson <ae@op5.se>