xrddefault: Fix custom variables persistence accros restarts
This fix is not Naemon specific, it is also applicable to Nagios.

Edit Sven Nierlein <Sven.Nierlein@consol.de>:

Naemon writes changed custom variables prefixed with an option
value into the retention.dat like

_VARNAME=1;value

On startup it then checks if the lenght of 1;value is more
than 3 which in fact skips changed custom variables if they
only contain a single character or nothing at all.