Remove rbtree and use GTree instead
The rbtree implementation was incorrect at least when removing objects
from the tree. Resulting in incorrect behavior. The following behaviour
did not always work:

while(!is_empty()) {
    remove( pick_one() );
}
assert( num_nodes() == 0 ); /* broke sometimes... */

Also, some parts of the implementation was made in #define:s, and some
in function calls, making the implementation really hard to modify.

Also, the API depended on direct access to structs.

Thus, throw away rbtree, and use something that is provided and
maintained from the outside. GTree in glib (we already depend on glib)
is therefore a much better choice.

This fixes MON-9066 and MON-7474.

Change-Id: I9caef9fc2e994ad7f96d257a213228cdcd6e6da3
Signed-off-by: Robin Hagman <robin.hagman@op5.com>
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>