Wait for command worker to terminate
This patch fixes a couple of issues with the handling of the "command
worker".

Previously, the command worker would be sent a SIGKILL on
shutdown - which only happens when it fails to read from its bufferqueue
- and hopefully have time to terminate before we tried to relaunch it.

On shutdown of the entire Naemon process, this would often trigger a lag
between the termination of the main process and the command file worker,
which would cause the command file worker to remain alive for a couple
of milliseconds. This, in turn, would sometimes cause issues when
trying to restart Naemon.

This patch changes the signal sent to the worker process to SIGTERM (and
resets the disposition of the same to the default handler in the
worker), and makes sure to wait for the worker to terminate before
returning from shutdown_command_file_worker().

In addition, we make sure to call the function before removing any lock
files to ensure that we can safely be restarted without having to worry
about lingering processes.

This fixes MON-8810.

Signed-off-by: Anton Lofgren <alofgren@op5.com>