neb: Add NEB callback API version 2
This new version of the NEB callback API provides NEB modules with the
ability to specify a reason for why they take a particular action. The
old API (now called API version 1) is still functional and remains
unchanged, which means that this patch should not break any existing
modules.

The new API, instead of emitting a single return code, returns a set of
"neb_cb_result"s, which in addition to the aforementioned description
and return code, also contains the module "name". The name is currently
simply the filename that we loaded the module from, or the string
"Unnamed core module", in the case of core modules. This could be
improved in the future to include a "pretty" name for modules, as well
as the ability to set names for core modules.

This patch also moves the existing NEB callback tests from t-tap to
libcheck, in order to make use of the latters excellent support for
fixtures, among other things.

This is part of the fix for MON-7564.

Signed-off-by: Anton Lofgren <alofgren@op5.com>