fix realtime clock chaos
first we mixed up glib and glibc, but it somehow still worked. It would
have worked with any library. The test program just checks if it compiles
and if the function exists. So checking for clock_gettime in glib worked
because it is either already in glibc or in none of them. However, we
had a check for clock_gettime already in place, it just didn't work
due to a typo. Removing the underline makes it work again.