signals: Zero out sigaction flags
Hoping that our uninitialized sigaction flags are magically set to a
"correct" value through the power of undefined behaviour is not a good
idea. In this particular case, the triggered behaviour was that sending
two SIGHUP signals in quick succession to the command file worker
caused it to terminate (even though we've specifically stated that we
want it to be handled by our sighandler!) - and since we don't handle
SIGCHLD, it would become a zombie.

This behaviour, which for the record is identical to the one favoured by
signal(2) which we recently moved away from, is caused by the sa_flags
fields being set to the respectable value, -767833534:

Breakpoint 1, setup_sighandler () at src/naemon/utils.c:246
(gdb) p sigact.sa_flags
$1 = -767833534

Or in binary:

(gdb) p /t sigact.sa_flags
$2 = 11010010001110111100101001000010

Which happens to correspond to (at least) the flags:

SA_RESETHAND (hex: 0x80000000, bin: 10000000000000000000000000000000)
which resets the signal disposition to SIG_DFL on entry to
handler, and

SA_NODEFER (hex: 0x40000000, bin: 1000000000000000000000000000000),
which causes delivery of a signal whose handler is currently being
executed to *not* be blocked.

This causes the command file worker to zombify since SIG_DFL for SIGHUP
is process termination, and we will not know to reap it.

This fixes MON-9322.

Signed-off-by: Anton Lofgren <alofgren@op5.com>