Respect retry intervals for service & host checks
The retry interval of a service or host should always be respected, even
in the case where it would cause the actual check to occur later than it
otherwise would (i.e when check interval < retry interval).

This patch adds a new check option, which explicitly allows scheduling
of checks that would make the check occur later than normal. We do this
rather than deleting the check event since we want to minimize (by
making the code easier to follow) the risk that we end up without any
check event at all.

This fixes MON-8745.

Signed-off-by: Anton Lofgren <alofgren@op5.com>