Correct logic in check_window macro
For certain commands that implicitly schedules a check on one or more
objects, we use the check_window macro to determine, for the affected
object type, whether we should schedule the next check according to the
retry interval, or the check interval.

However, this macro contained a bug which would cause it to only use the
retry interval for SOFT OK/UP states. This is in contrast with the
regular logic for when we use the retry interval. Namely, we use the
retry interval in SOFT *non*-OK/UP states.

This patch fixes that.

This, together with the preceding patch fixes Github issue #117.

Signed-off-by: Anton Lofgren <alofgren@op5.com>