Let configure warn if `help2man` is not present on the system
We don't want to halt the build entirely if it's not present, however,
since it is not necessary for all use cases.

In particular, it is usually not necessary for a contributor to be able
to generate man pages. Rather, this should be done by someone who
intends to create a distribution.

This is related to Github issue #95.

Signed-off-by: Anton Lofgren <alofgren@op5.com>