Handle overflow when diffing times
The tv_sec field in a timespec struct is defined as a time_t, which is
often a long int (but not always). On 64-bit systems this is usually
fine (except when it's not), but on 32-bit systems, this can have the
consequence of the arithmetic in the timespec_msdiff function
overflowing the resulting diff, for example when given a large negative
value such as (0 - time(NULL)). This in turn would lead to a diff a
number of years in the future.

For example, (time_t)-1464166432 * 1000 = 417415936 which is about 13
years in the future.

This has a nasty consequence, because of how we use this function, in that it would

a) cause events scheduled way back to be scheduled over a decade too
late, and
b) thusly block the execution of *all other events* until that event was
run.

This patch fixes that issue, by defining what happens in the case of an
overflow more strictly, while still maintaining the desired properties.
Namely, we simply return LONG_MIN/LONG_MAX if event A is too far in the
past or future relative to event B.

In addition, I also added a bunch of tests to catch this type of bug in
the future, as well as portable support for overflow-checked arithmetics.

Signed-off-by: Anton Lofgren <alofgren@op5.com>