commands: Don't support undocumented arguments
For some inexplicable reason, the RESTART_PROGRAM and SHUTDOWN_PROGRAM
commands (along with their *_PROCESS aliases) grew an entry_time
parameter somewhere along the line.

This has never been documented or commented upon, and since I was the
one to add them in the first case, according to git, I take the
responsibility of removing them as well.

We found this, since when they are *not* set, they would cause an
event to be scheduled a long time in the past instead of right now,
which triggered a bug in the event handling loop.

Signed-off-by: Anton Lofgren <alofgren@op5.com>