iobroker: Allow polling without registered FDs
Returning an error when we have no reigstered file descriptors to listen
on is slightly inconsistent with the way we use this function. That is,
as a function that either returns the number of events available OR
idles for a specified number of milliseconds. The natural way to behave
for a function like this - which is also the behaviour of select(),poll()
and epoll() - is to simply accept that no events will arrive and idle
for the duration of the timeout.

Signed-off-by: Anton Lofgren <alofgren@op5.com>