worker: Don't reap grandchildren immediately
In order to provide backwards compatibility with older configurations,
this patch lends some leniency to when grandchild processes are killed.

It used to be the case that a process group was killed as soon as we
reaped its leader. With this patch, we instead allow the descendants of
the leader to continue running until the timeout of the job in question
has been reached, at which point all still remaining processes will be
reaped.

This allows users to continue using, for example, asynchronous mail
delivery via mailx, without losing their notifications when mailx's
children are unexpectedly killed.

In addition, this patch also cleans up the worker test suite a bit, to
make it more reliable and extendable.

This fixes GitHub issue #137.

This is related to MON-8090.

Signed-off-by: Anton Lofgren <alofgren@op5.com>