Worker: Don't pass a callback through spawning
The callback was probably meant to make it possible to create a custom
worker with a separate job spawning method. But since all those methods
were local to that file, no external builds could have used that method.
In fact, the struct for tracking jobs didn't provide any space for
custom job related data, thus this wasn't even possible.

Instead of fixing it without a use case, it's just better to remove, to
simplify the debugging of the code.

Part of op5 issue MON-9528

Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>