Worker: Mark as stale when finishing jobs
The ESTALE seems to be when the job has any processes left that needs to
be reaped. That's what happens when the main process finishes, and we
call finish_job.

Thus, mark the job as ESTALE, and don't log when we kill dormant
children, since that's the default behaviour for jobs, where we still
need to be able to have dangling processes a while after execution.

Those dangling processes we have had is for example sendmail in
notification scripts, which forks a process to send the mail and exit
early.

Resolves: MON-9528

Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>