fix worker writing to already rotated logfile
If a worker exits because the main naemon process has exited it writes
one last log entry. This entry is likely to end up in an already rotated
old logfile which breaks livestatus.
It then exits with EXIT_SUCCESS, so its not really an error anyway and
therefor we can just lower the loglevel.

Signed-off-by: Sven Nierlein <sven@nierlein.de>