Revert "Merge pull request #52 in MONITOR/naemon from feature/MON-10198-remove-unnecessary-configuration to master"
Since naemon starts with daemonize option under EL6, we fork and report the
success of the fork, not the actual return code of naemon started.

So if we have a config change that broke our config, the exit would still
be 0, but the service then fail at config parse and die a second later.

This reverts commit 086d1cb16f64d256457dd51a07a17c14cd9f8bbd, reversing
changes made to 4e83227d6b55cd57f32bf79fdd60433e3e52d2be.