naemon.c: signal parent upon daemonization
Previously, starting naemon as a daemon would return an OK return code
even if the object configuration was bad. The reason being that the
configuration parsing is done by the forked child and the parent exits
as soon as it has forked okay.

This patch makes the parent process wait for the child to complete load
everything up until it is about to enter the main event loop. By listening
to an unnamned pipe for input from the child we will block until we receive
input. If the child were to crash during initialization we will handle the
SIGCHLD signal and exit the parent process with a bad exit code. That way
we will not end up in a state where we're blocking forever because the
child has died.

So now we will get a bad exit code when daemonizing if the configuration
is bad or if the program dies for any reason during initialization.

Part of MON-10198

Signed-off-by: Robin Engström <robin.engstrom@op5.com>