daemon-init.in: don't verify config when starting daemon
Don't verify config when starting naemon via systemv so that we don't
bail out in case of a faulty configuration. Instead we use the return
code of naemon to display how initialization went.

The same goes for restarting. If we want to verify the configuration
it is still possible using the -v flag.

Part of MON-10198

Signed-off-by: Robin Engström <robin.engstrom@op5.com>