utils.c: don't use SIGCHLD to signal daemon child exit
In a previous commit a bug was introduced caused by the handling of SIGCHLD.
If the signal was received we would do exit(EXIT_FAILURE) which was okay for the
original process.

However the problem occured when workers processes of the daemon child were exiting.
It would then cause the process to instantaeously die without doing cleanup of any kind.

With this patch we instead rely on the unnamned pipe used to communicate between the
original process and the daemon child when daemonizing. If the child dies the parent
will stop blocking and exit with a failure unless the child says otherwise.

Part of MON-10198

Signed-off-by: Robin Engström <robin.engstrom@op5.com>