lower wproc: logging log level
Do not log timeouts 3 times for each timeout. Timeouts are a normal thing in
monitoring and state changes are logged already, there is no need to add 2
extra log entries, one from wproc and one from the check handler. Especially since
the timeout is quite visible already and the extra logging does not add any value.
Since this is used in tests, we have to change the tests to check the debug logfile
for the expected output.

Signed-off-by: Sven Nierlein <sven@nierlein.de>